Refactored mapbox map
