+++ src/features/projects/components/MapboxMap.tsx
@@ -1,5 +1,3 @@
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormGroup from '@material-ui/core/FormGroup';
 /* eslint-disable no-nested-ternary */
 /* eslint-disable no-underscore-dangle */
 import * as turf from '@turf/turf';
@@ -15,30 +13,13 @@ import MapGL, {
   Source,
   WebMercatorViewport,
 } from 'react-map-gl';
-import { LayerManager, Layer as LayerM } from 'layer-manager/dist/components';
-import { PluginMapboxGl } from 'layer-manager';
-import CancelIcon from '../../../../public/assets/images/icons/CancelIcon';
-import ExploreIcon from '../../../assets/images/icons/ExploreIcon';
-import Switch from '../../common/InputTypes/ToggleSwitch';
 import LeftIcon from '../../../../public/assets/images/icons/LeftIcon';
 import RightIcon from '../../../../public/assets/images/icons/RightIcon';
-import { getParams } from '../../../utils/LayerManagerUtils';
-import TreeCoverLoss from '../../../../public/data/layers/tree-cover-loss';
-import {
-  Icons,
-  Legend,
-  LegendListItem,
-  LegendItemTimeStep,
-} from 'vizzuality-components';
-
 import styles from '../styles/MapboxMap.module.scss';
-import InfoIcon from '../../../../public/assets/images/icons/InfoIcon';
-import OpenLink from '../../../../public/assets/images/icons/OpenLink';
-import i18next from '../../../../i18n/';
 import { Modal } from '@material-ui/core';
-import ProjectSnippet from './ProjectSnippet';
-
-const { useTranslation } = i18next;
+import ExploreInfoModal from './maps/ExploreInfoModal';
+import ExploreContainer from './maps/ExploreContainer';
+import PopupProject from './PopupProject';
 
 interface mapProps {
   projects: any;
@@ -62,8 +43,6 @@ export default function MapboxMap({
   let timer: NodeJS.Timeout;
   const router = useRouter();
 
-  const { t, i18n } = useTranslation(['maps']);
-
   const mapRef = useRef(null);
   const exploreContainerRef = useRef(null);
   const screenWidth = window.innerWidth;
@@ -83,7 +62,6 @@ export default function MapboxMap({
   const buttonRef = useRef(null);
   const popupRef = useRef(null);
   const [loaded, setLoaded] = useState(false);
-  const [layersSettings, setLayersSettings] = useState({});
   const infoRef = useRef(null);
 
   const [mapState, setMapState] = useState({
@@ -101,14 +79,6 @@ export default function MapboxMap({
 
   const [exploreExpanded, setExploreExpanded] = React.useState(false);
 
-  const [exploreForests, setExploreForests] = React.useState(false);
-
-  const [explorePotential, setExplorePotential] = React.useState(false);
-
-  const [exploreDeforestation, setExploreDeforestation] = React.useState(false);
-
-  const [explorePlanted, setExplorePlanted] = React.useState(false);
-
   const [exploreProjects, setExploreProjects] = React.useState(true);
 
   const [infoExpanded, setInfoExpanded] = React.useState(null);
@@ -121,19 +91,6 @@ export default function MapboxMap({
     setModalOpen(true);
   };
 
-  const handleExploreForestsChange = (event) => {
-    setExploreForests(event.target.checked);
-  };
-
-  const handleExplorePotentialChange = (event) => {
-    setExplorePotential(event.target.checked);
-  };
-  const handleExploreDeforestationChange = (event) => {
-    setExploreDeforestation(event.target.checked);
-  };
-  const handleExplorePlantedChange = (event) => {
-    setExplorePlanted(event.target.checked);
-  };
   const handleExploreProjectsChange = (event) => {
     setExploreProjects(event.target.checked);
     setShowProjects(event.target.checked);
@@ -316,11 +273,11 @@ export default function MapboxMap({
 
   const _onViewportChange = (view: any) => setViewPort({ ...view });
 
-  const handleClose = () => {
-    setOpen(false);
-  };
   const handleOpen = () => {
     setOpen(true);
+  };
+  const handleClose = () => {
+    setOpen(false);
   };
 
   function goToNextProject() {
@@ -338,60 +295,6 @@ export default function MapboxMap({
       setCurrentSite(maxSites - 1);
     }
   }
-
-  // LEGEND
-  const layerLegend = TreeCoverLoss.map((l) => {
-    const { id, decodeConfig, timelineConfig } = l;
-    const lSettings = layersSettings[id] || {};
-
-    const decodeParams =
-      !!decodeConfig &&
-      getParams(decodeConfig, { ...timelineConfig, ...lSettings.decodeParams });
-    const timelineParams = !!timelineConfig && {
-      ...timelineConfig,
-      ...getParams(decodeConfig, lSettings.decodeParams),
-    };
-
-    return {
-      id,
-      slug: id,
-      dataset: id,
-      layers: [
-        {
-          active: true,
-          ...l,
-          ...lSettings,
-          decodeParams,
-          timelineParams,
-        },
-      ],
-      ...lSettings,
-    };
-  });
-
-  const onChangeLayerDate = (dates, layer) => {
-    const { id, decodeConfig } = layer;
-
-    setLayersSettings({
-      ...layersSettings,
-      [id]: {
-        ...layersSettings[id],
-        ...(decodeConfig && {
-          decodeParams: {
-            startDate: dates[0],
-            endDate: dates[1],
-            trimEndDate: dates[2],
-          },
-        }),
-        ...(!decodeConfig && {
-          params: {
-            startDate: dates[0],
-            endDate: dates[1],
-          },
-        }),
-      },
-    });
-  };
 
   return (
     <div className={styles.mapContainer}>
@@ -485,12 +388,7 @@ export default function MapboxMap({
                   clearTimeout(timer);
                 }}
                 onFocus={() => {}}
-              >
-                {/* <img
-                src="https://cdn-app.plant-for-the-planet.org/media/maps/pet_p.svg"
-                className={styles.markerType}
-              /> */}
-              </div>
+              />
             </Marker>
           ))}
         {popupData.show && !isMobile && (
@@ -544,244 +442,35 @@ export default function MapboxMap({
                 }
               }}
             >
-              <ProjectSnippet 
-                key={popupData.project.properties.id}
-                project={popupData.project.properties}
-                editMode={false}
+              <PopupProject	              
+                key={popupData.project.properties.id}	              
+                project={popupData.project}	             
+                buttonRef={buttonRef}	             
+                popupRef={popupRef}	     
+                open={open}	 
+                handleOpen={handleOpen}	
+                handleClose={handleClose}	
               />
             </div>
           </Popup>
         )}
-        {exploreForests ? (
-          <Source
-            id="forests"
-            type="raster"
-            tiles={[
-              'https://tiles.arcgis.com/tiles/lKUTwQ0dhJzktt4g/arcgis/rest/services/Forest_Denisty_V2/MapServer/tile/{z}/{y}/{x}',
-            ]}
-            tileSize={128}
-          >
-            <Layer id="forest-layer" source="forests" type="raster" />
-          </Source>
-        ) : null}
-
-        {/* {explorePotential ? (
-          <Source
-            id="potential"
-            type="raster"
-            tiles={[
-              'https://tiles.arcgis.com/tiles/lKUTwQ0dhJzktt4g/arcgis/rest/services/WWF_Restoration_V3/MapServer/tile/{z}/{y}/{x}',
-            ]}
-            tileSize={128}
-          >
-            <Layer id="potential-layer" source="potential" type="raster" />
-          </Source>
-        ) : null} */}
-
-        {loaded ? (
-          <LayerManager map={mapRef.current.getMap()} plugin={PluginMapboxGl}>
-            {exploreDeforestation &&
-              TreeCoverLoss.map((layer) => {
-                const {
-                  id,
-                  decodeConfig,
-                  timelineConfig,
-                  decodeFunction,
-                } = layer;
-
-                const lSettings = layersSettings[id] || {};
-
-                const l = {
-                  ...layer,
-                  ...layer.config,
-                  ...lSettings,
-                  ...(!!decodeConfig && {
-                    decodeParams: getParams(decodeConfig, {
-                      ...timelineConfig,
-                      ...lSettings.decodeParams,
-                    }),
-                    decodeFunction,
-                  }),
-                };
-
-                return <LayerM key={layer.id} {...l} />;
-              })}
-          </LayerManager>
-        ) : null}
 
         <div className={styles.mapNavigation}>
           <NavigationControl showCompass={false} />
         </div>
-        <div ref={exploreContainerRef}>
-          <div
-            className={styles.exploreButton}
-            onClick={() => {
-              if (exploreExpanded) {
-                setExploreExpanded(false);
-              } else {
-                setExploreExpanded(true);
-              }
-            }}
-            style={
-              exploreExpanded
-                ? {
-                    padding: '4px 10px',
-                  }
-                : {}
-            }
-          >
-            {exploreExpanded ? <CancelIcon /> : <ExploreIcon />}
-            {exploreExpanded ? null : (
-              <p
-                onClick={() => setExploreExpanded(true)}
-                className={styles.exploreText}
-              >
-                {isMobile ? null : t('maps:explore')}
-              </p>
-            )}
-          </div>
-          {exploreExpanded ? (
-            <>
-              <div className={styles.exploreExpanded}>
-                {/* <div> */}
-                <FormGroup style={{ width: '100%' }}>
-                  <div className={styles.exploreToggleRow}>
-                    <FormControlLabel
-                      control={
-                        <Switch
-                          color="#448149"
-                          className={styles.toggleForest}
-                          checked={exploreForests}
-                          onChange={handleExploreForestsChange}
-                          name="forest"
-                        />
-                      }
-                      label={t('maps:forests')}
-                    />
-                    <div
-                      onClick={() => {
-                        setInfoExpanded('Forests');
-                        setModalOpen(true);
-                      }}
-                      className={styles.exploreInfo}
-                    >
-                      <InfoIcon />
-                    </div>
-                  </div>
-                  {/* <div className={styles.exploreToggleRow}>
-                      <FormControlLabel
-                        control={
-                          <Switch
-                            checked={explorePotential}
-                            onChange={handleExplorePotentialChange}
-                            name="potential"
-                          />
-                        }
-                        label={t('maps:restoration')}
-                      />
-                      <div
-                        onClick={() => {
-                          setInfoExpanded('Restoration');
-                        }}
-                        className={styles.exploreInfo}
-                      >
-                        <InfoIcon />
-                      </div>
-                    </div> */}
 
-                  <div className={styles.exploreToggleRow}>
-                    <FormControlLabel
-                      control={
-                        <Switch
-                          color="#FF0000"
-                          checked={exploreDeforestation}
-                          onChange={handleExploreDeforestationChange}
-                          name="deforestation"
-                        />
-                      }
-                      label={t('maps:deforestation')}
-                    />
-                    <div
-                      onClick={() => {
-                        setInfoExpanded('Deforestation');
-                        setModalOpen(true);
-                      }}
-                      className={styles.exploreInfo}
-                    >
-                      <InfoIcon />
-                    </div>
-                  </div>
-                  {exploreDeforestation ? (
-                    <div className={styles.deforestionSlider}>
-                      <Icons />
-                      <Legend collapsable={false} sortable={false}>
-                        {layerLegend.map((layerGroup, i) => {
-                          return (
-                            <LegendListItem
-                              index={i}
-                              key={layerGroup.slug}
-                              layerGroup={layerGroup}
-                              className={styles.layerLegend}
-                            >
-                              {/* <LegendItemTypes /> */}
-                              <LegendItemTimeStep
-                                defaultStyles={{
-                                  handleStyle: {
-                                    backgroundColor: 'white',
-                                    borderRadius: '50%',
-                                    boxShadow:
-                                      '0 1px 2px 0 rgba(0, 0, 0, 0.29)',
-                                    border: '0px',
-                                    zIndex: 2,
-                                  },
-                                  railStyle: { backgroundColor: '#d6d6d9' },
-                                  dotStyle: {
-                                    visibility: 'hidden',
-                                    border: '0px',
-                                  },
-                                }}
-                                handleChange={onChangeLayerDate}
-                              />
-                            </LegendListItem>
-                          );
-                        })}
-                      </Legend>
-                    </div>
-                  ) : null}
-                  {/* <div className={styles.exploreToggleRow}>
-                    <FormControlLabel
-                      control={
-                        <Switch
-                         color="#E7C746"
-                          checked={explorePlanted}
-                          onChange={handleExplorePlantedChange}
-                          name="planted"
-                        />
-                      }
-                      label="Planted Trees"
-                    />
-                  </div> */}
-                  <div className={styles.exploreToggleRow}>
-                    <FormControlLabel
-                      control={
-                        <Switch
-                          checked={exploreProjects}
-                          onChange={handleExploreProjectsChange}
-                          name="projects"
-                        />
-                      }
-                      label={t('maps:projects')}
-                    />
-                  </div>
-                </FormGroup>
-                {/* </div> */}
-                <div className={styles.exploreCaption}>
-                  <p>{t('maps:3trilliontrees')}</p>
-                </div>
-              </div>
-            </>
-          ) : null}
-        </div>
+        <ExploreContainer 
+          exploreContainerRef={exploreContainerRef} 
+          setExploreExpanded={setExploreExpanded} 
+          exploreExpanded={exploreExpanded} 
+          isMobile={isMobile} 
+          setInfoExpanded={setInfoExpanded} 
+          setModalOpen={setModalOpen} 
+          loaded={loaded}
+          mapRef={mapRef}
+          handleExploreProjectsChange={handleExploreProjectsChange}
+          exploreProjects={exploreProjects}
+        />
 
         {showSingleProject && siteExists ? (
           maxSites! > 1 ? (
@@ -825,92 +514,9 @@ export default function MapboxMap({
           aria-labelledby="simple-modal-title"
           aria-describedby="simple-modal-description"
         >
-          <div ref={infoRef} className={styles.infoExpanded}>
-            {infoExpanded === 'Forests' ? (
-              <div className={styles.infoContainer}>
-                <div className={styles.infoTitle}>{t('maps:forests')}</div>
-                <div className={styles.infoContent}>
-                  <div className={styles.currentForestScale}>
-                    <p>{t('maps:low')}</p>
-                    <div></div>
-                    <p>{t('maps:high')}</p>
-                  </div>
-                  <p>{t('maps:forestInfo')}</p>
-                  <a
-                    href="https://www.nature.com/articles/nature14967"
-                    target="_blank"
-                    style={{ paddingTop: 20 }}
-                  >
-                    <OpenLink />
-                    <p>
-                      Crowther, T. W. et al. (2015) Mapping tree
-                      <br /> density at a global scale. Nature 525, 201–205.
-                    </p>
-                  </a>
-                </div>
-              </div>
-            ) : null}
-            {infoExpanded === 'Restoration' ? (
-              <div className={styles.infoContainer}>
-                <div className={styles.infoTitle}>{t('maps:restoration')}</div>
-                <div className={styles.infoContent}>
-                  <div className={styles.reforestationScale}>
-                    <p>{t('maps:low')}</p>
-                    <div></div>
-                    <p>{t('maps:high')}</p>
-                  </div>
-                  <p>{t('maps:restorationInfo')}</p>
-                  <a
-                    href="https://science.sciencemag.org/content/365/6448/76"
-                    target="_blank"
-                    style={{ paddingTop: 20 }}
-                  >
-                    <OpenLink />
-                    <p>
-                      Bastin, J. F. et al. (2019) The Global Tree
-                      <br /> Restoration Potential. Science 365(6448), 76-79.
-                    </p>
-                  </a>
-                </div>
-              </div>
-            ) : null}
-            {infoExpanded === 'Deforestation' ? (
-              <div className={styles.infoContainer}>
-                <div className={styles.infoTitle}>
-                  {t('maps:deforestation')}
-                </div>
-                <div className={styles.infoContent}>
-                  <a
-                    href="https://data.globalforestwatch.org/datasets/63f9425c45404c36a23495ed7bef1314"
-                    target="_blank"
-                    style={{ paddingTop: 20 }}
-                  >
-                    <OpenLink />
-                    <p>
-                      Global Forest Watch
-                      <br />
-                      globalforestwatch.org
-                    </p>
-                  </a>
-                </div>
-              </div>
-            ) : null}
-            {infoExpanded === 'Planted' ? (
-              <div className={styles.infoContainer}>
-                <div className={styles.infoTitle}>{infoExpanded}</div>
-                <div className={styles.infoContent}></div>
-              </div>
-            ) : null}
-            <div
-              onClick={() => {
-                setInfoExpanded(null);
-                setModalOpen(false);
-              }}
-              className={styles.infoClose}
-            >
-              <CancelIcon color="#d5d5d5" />
-            </div>
-          </div>
+          <ExploreInfoModal 
+            infoRef={infoRef} infoExpanded={infoExpanded} setInfoExpanded={setInfoExpanded} setModalOpen={setModalOpen} 
+          />
         </Modal>
       ) : null}
     </div>
